requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.4.0
cryptography==42.0.3
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==4.2.10
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-nested-routers==0.93.5
exceptiongroup==1.2.0
flower==2.0.1
gevent==24.2.1
greenlet==3.0.3
gunicorn==21.2.0
humanize==4.9.0
idna==3.6
iniconfig==2.0.0
kafka-python==2.0.2
kombu==5.3.5
model-bakery==1.17.0
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
pluggy==1.3.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pytest==7.4.4
pytest-cov==4.1.0
pytest-django==4.7.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
tomli==2.0.1
tornado==6.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
zope.event==5.0
zope.interface==6.2
